Naver_Posting-main/requirements_streamlit.txt
# 네이버 자동 포스팅 시스템 - 필수 라이브러리

# Streamlit
streamlit==1.29.0

# 데이터 처리
pandas==2.1.4
numpy==1.24.3

# 이미지 처리
Pillow==10.1.0

# Selenium 웹 자동화
selenium==4.16.0
webdriver-manager==4.0.1

# 클립보드 조작
clipboard==0.0.4

# Google Gemini AI
google-generativeai==0.3.2

# HTTP 요청
requests==2.31.0

# 기타 유틸리티
python-dotenv==1.0.0